% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the sensory database}
\usage{
connect_db(dbdir = app_sys("database", "sensory_db.sqlite"), read_only = TRUE)
}
\arguments{
\item{dbdir}{Character. Path to the sensory database.}

\item{read_only}{Logical. Define a read only connection or not.}
}
\value{
The connection to the database.
}
\description{
Connect to the sensory database
}
\examples{
con_db <- connect_db()

# List tables in the sensory database
DBI::dbListTables(con_db)

# Read the sensory profiles
dplyr::tbl(con_db, dplyr::sql("SELECT * FROM PROFILES LIMIT 10"))

# Read the hedonic data
dplyr::tbl(con_db, dplyr::sql("SELECT * FROM HEDONIC LIMIT 10"))

# Read the sessions data
dplyr::tbl(con_db, dplyr::sql("SELECT * FROM SESSIONS"))

# Read the products data
dplyr::tbl(con_db, dplyr::sql("SELECT * FROM PRODUCTS LIMIT 10"))

# Disconnect
DBI::dbDisconnect(con_db, shutdown = TRUE)
}
