% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create the reproducible HTML sensory report}
\usage{
create_report(
  author,
  session,
  products,
  output_dir,
  report_template = app_sys("report", "report_template.qmd")
)
}
\arguments{
\item{author}{Character. Author of the report.}

\item{session}{Character. ID of the sensory session.}

\item{products}{Character. Vector of products to include into the report.}

\item{output_dir}{Character. Path of the folder to save the report.}

\item{report_template}{Character. Path of the qmd report template.}
}
\value{
A list with the folder and the file of the created report.
}
\description{
Create the reproducible HTML sensory report
}
\examples{
# Create a temporary directory
my_temp_dir <- tempfile()
dir.create(my_temp_dir)

# Create the sensory report
res_info_report <- create_report(
  author = "Swann",
  session = "SESS2301",
  products = c("PROD1", "PROD2", 
               "PROD3", "PROD4", 
               "PROD5", "PROD6"),
  output_dir = my_temp_dir
)

# Open the sensory report
browseURL(
  file.path(
    res_info_report$output_dir_report, 
    res_info_report$output_file_report
  )
)

# Delete temporary directory
unlink(my_temp_dir, recursive = TRUE)
}
