% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_coord_real_toy.R
\docType{data}
\name{data_coord_real_toy}
\alias{data_coord_real_toy}
\title{data_coord_real_toy}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{ dim1 }{  numeric }
\item{ dim2 }{  numeric }
\item{ dim1dim2 }{  matrix,array }
}
}
\source{
Source
}
\usage{
data_coord_real_toy
}
\description{
Description.
}
\keyword{datasets}
