% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_hedonic_toy.R
\docType{data}
\name{data_hedonic_toy}
\alias{data_hedonic_toy}
\title{data_hedonic_toy}
\format{
A data frame with 400 rows and 3 variables:
\describe{
\item{ PRODUCT }{  character }
\item{ CONSUMER }{  character }
\item{ SCORE }{  numeric }
}
}
\source{
Source
}
\usage{
data_hedonic_toy
}
\description{
Description.
}
\keyword{datasets}
