% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_products_toy.R
\docType{data}
\name{data_products_toy}
\alias{data_products_toy}
\title{data_products_toy}
\format{
A data frame with 16 rows and 3 variables:
\describe{
\item{ PRODUCT }{  character }
\item{ INFO }{  character }
\item{ BRAND }{  character }
}
}
\source{
Source
}
\usage{
data_products_toy
}
\description{
Description.
}
\keyword{datasets}
