% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_profiles_toy.R
\docType{data}
\name{data_profiles_toy}
\alias{data_profiles_toy}
\title{data_profiles_toy}
\format{
A data frame with 52 rows and 3 variables:
\describe{
\item{ PRODUCT }{  character }
\item{ ATTRIBUTE }{  character }
\item{ SCORE }{  numeric }
}
}
\source{
Source
}
\usage{
data_profiles_toy
}
\description{
Description.
}
\keyword{datasets}
