% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_sensory_toy.R
\docType{data}
\name{data_sensory_toy}
\alias{data_sensory_toy}
\title{data_sensory_toy}
\format{
A data frame with 160 rows and 3 variables:
\describe{
\item{ PRODUCT }{  character }
\item{ CONSUMER }{  character }
\item{ GROUP }{  character }
}
}
\source{
Source
}
\usage{
data_sensory_toy
}
\description{
Description.
}
\keyword{datasets}
