% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_sessions_toy.R
\docType{data}
\name{data_sessions_toy}
\alias{data_sessions_toy}
\title{data_sessions_toy}
\format{
A data frame with 6 rows and 5 variables:
\describe{
\item{ SESSION }{  character }
\item{ DATE }{  character }
\item{ DEPARTMENT }{  character }
\item{ COUNTRY }{  character }
\item{ LEADER }{  character }
}
}
\source{
Source
}
\usage{
data_sessions_toy
}
\description{
Description.
}
\keyword{datasets}
