% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_prefmap.R
\name{perform_prefmap}
\alias{perform_prefmap}
\title{Perform the preference mapping}
\usage{
perform_prefmap(
  res_mapping,
  data_hedonic,
  color_min = "white",
  color_max = "#55B4D2",
  resolution = 200
)
}
\arguments{
\item{res_mapping}{List. Result of the perform_senso_mapping() function.}

\item{data_hedonic}{Dataframe. Hedonic data.}

\item{color_min}{Character. Color for "do not like".}

\item{color_max}{Character. Color for "like".}

\item{resolution}{Integer. Resolution for the interactive plot.}
}
\value{
A plotly preference mapping plot
}
\description{
Perform the preference mapping
}
\examples{
data("data_hedonic_toy")

res_mapping <- perform_senso_mapping(
  data_profiles = data_profiles_toy,
  data_products = data_products_toy
)

perform_prefmap(
  res_mapping,
  data_hedonic_toy
)
}
