% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unidim_analysis.R
\name{plot_unidim_analysis}
\alias{plot_unidim_analysis}
\title{Create the plot with the unidimensional analysis}
\usage{
plot_unidim_analysis(data_sensory)
}
\arguments{
\item{data_sensory}{Tibble. The sensory data.}
}
\value{
A ggplotly
}
\description{
Create the plot with the unidimensional analysis
}
\examples{
data("data_sensory_toy")

plot_unidim_analysis(
  data_sensory = data_sensory_toy
)
}
