% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensoreport-package.R
\docType{package}
\name{sensoreport-package}
\alias{sensoreport}
\alias{sensoreport-package}
\title{sensoreport: Create Reproducible Sensory Reports}
\description{
A Shiny app to query a database in which QDA sensory data are stored, analyse data R package and generate reproducible and interactive HTML sensory reports. Such tool can enable sensory analysts and panel leaders to spend less time preparing reports, and add more value to their work.
}
\author{
\strong{Maintainer}: Margot Brard \email{margot@thinkr.fr} (\href{https://orcid.org/0000-0001-6754-0659}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
