% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the sensory database}
\usage{
connect_db(dbdir = app_sys("database", "sensory_db.duckdb"))
}
\arguments{
\item{dbdir}{Character. Path to the sensory database.}
}
\value{
The connection to the database.
}
\description{
Connect to the sensory database
}
\examples{
con_db <- connect_db()

# List tables in the sensory database
DBI::dbListTables(con_db)

# Read the sensory profiles
# dplyr::tbl(con_db, "PROFILES")
  
# Disconnect
DBI::dbDisconnect(con_db, shutdown = TRUE)
}
